+++ examples/metadata.rs
@@ -2,7 +2,7 @@ fn main() {
     use gdal::{Dataset, Metadata};
     use std::path::Path;
 
-    let driver = gdal::Driver::get("mem").unwrap();
+    let driver = gdal::Driver::get_by_name("mem").unwrap();
     println!("driver description: {:?}", driver.description());
 
     let path = Path::new("./fixtures/tinymarble.png");