+++ src/programs/raster/mod.rs
@@ -1,0 +1,2 @@
+mod vrt;
+pub use vrt::*;